# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TN Bus+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buses/{id}:
    get:
      operationId: getBus
      tags: [buses]
      summary: Get bus by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bus details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
        "404":
          description: Not found

  # Search
...
  /schedules/{scheduleId}/seats:
    get:
      operationId: getSeats
      tags: [buses]
      summary: Get seat layout for a schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seat layout
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seat"

  # Bookings
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for a passenger
      parameters:
        - name: passengerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /pnr/{pnr}:
    get:
      operationId: getPnrStatus
      tags: [bookings]
      summary: Get PNR status
      parameters:
        - name: pnr
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNR status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found

  # Payments (Razorpay)
...
  /bookings/{id}/rebooking-suggestions:
    get:
      operationId: getRebookingSuggestions
      tags: [bookings]
      summary: Get smart rebooking suggestions for the same route as a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rebooking suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RebookingSuggestion"

  # Passenger booking preferences (seat / bus type memory)
...
components:
  schemas:
...
    Bus:
      type: object
      properties:
        id:
          type: integer
        busNumber:
          type: string
        busType:
          type: string
        totalSeats:
          type: integer
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
        currentLocation:
          type: ["string", "null"]
        punctualityScore:
          type: number
        driverName:
          type: ["string", "null"]
        routeId:
          type: ["integer", "null"]
      required:
        - id
        - busNumber
        - busType
        - totalSeats
        - status
...
    Seat:
      type: object
      properties:
        seatNumber:
          type: string
        row:
          type: integer
        column:
          type: integer
        type:
          type: string
        isAvailable:
          type: boolean
        isWomenOnly:
          type: boolean
        price:
          type: number
      required:
        - seatNumber
        - row
        - column
        - type
        - isAvailable
        - isWomenOnly
        - price
    Booking:
      type: object
      properties:
        id:
          type: integer
        pnr:
          type: string
        passengerId:
          type: integer
        scheduleId:
          type: integer
        busNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        seatNumbers:
          type: array
          items:
            type: string
        totalFare:
          type: number
        status:
          type: string
        paymentStatus:
          type: string
        passengerName:
          type: string
        passengerPhone:
          type: string
        createdAt:
          type: string
        qrCode:
          type: ["string", "null"]
        rewardPointsEarned:
          type: integer
          description: Reward points credited for this booking, derived from the trusted schedule fare. Present on the create-booking response and on booking detail.
      required:
        - id
        - pnr
        - passengerId
        - scheduleId
        - origin
        - destination
        - departureTime
        - arrivalTime
        - seatNumbers
        - totalFare
        - status
        - paymentStatus
        - passengerName
        - passengerPhone
        - createdAt
    BookingInput:
      type: object
      properties:
        passengerId:
          type: integer
        scheduleId:
          type: integer
        seatNumbers:
          type: array
          items:
            type: string
        passengerName:
          type: string
        passengerPhone:
          type: string
        totalFare:
          type: number
        paymentMethod:
          type: string
          description: How the fare is paid. When "wallet", the trusted server-side fare is debited from the passenger's wallet balance (this is how redeemed reward points are spent against fares). Any other value is treated as a simulated external payment.
      required:
        - passengerId
        - scheduleId
        - seatNumbers
        - passengerName
        - passengerPhone
        - totalFare
...
    RebookingSuggestion:
      type: object
      properties:
        scheduleId:
          type: integer
        busId:
          type: integer
        busNumber:
          type: string
        busType:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        fare:
          type: number
        availableSeats:
          type: integer
      required:
        - scheduleId
        - busId
        - busNumber
        - busType
        - origin
        - destination
        - departureTime
        - arrivalTime
        - fare
        - availableSeats
...